+++ ChangeLog.md
@@ -8,6 +8,7 @@
 * Fix exception when attempting to display ROI overlays
 * Add ability to extract embedded overlays from pixel data
 * Detect incorrect transfer syntax in file meta info
+* Add support for reading and displaying GE Private Implicit VR Big Endian syntax
 
 #### v1.0.23 (09/26/2012)
 * Fix W/L calculation creating negative window width